+++ spec/legion/extensions/transformer/runners/transform_schema_spec.rb
@@ -4,18 +4,6 @@ require 'spec_helper'
 require 'tilt'
 require 'json'
 
-# Stub framework helpers (guard against re-definition from other specs)
-unless defined?(Legion::Extensions::Helpers::Lex)
-  module Legion
-    module Extensions
-      module Helpers
-        module Lex; end
-        module Task; end
-      end
-    end
-  end
-end
-
 require 'legion/extensions/transformer/runners/transform'
 
 RSpec.describe Legion::Extensions::Transformer::Runners::Transform, '#transform with schema' do